Allow complete inlining of special member function calls
This gives the optimizer enough info to devirtualize calls to the
traits, and thus the special member functions of the implementation when
the implementation is a complete type at the time that these
trait-functions are called.

This causes smaller code, with less branches and less indirect calls, to
be generated when moving the special member function instantiations to
the same compilation unit as that containing the full definition of the
'implementation' type.